Feat(Assets): Assets and initialization
