+++ src/pages/Login/Login.tsx
@@ -1,0 +1,5 @@
+export default function Login(){
+    return(
+        <h1>login page goes here</h1>
+    )
+}
